+++ includes/widgets/mybooking-plugin-rent-selector-widget.php
@@ -44,6 +44,13 @@ class MyBookingRentEngineSelectorWidget extends WP_Widget {
             $data['rental_location_code'] = '';
         }
 
+        if ( array_key_exists('layout', $instance) ) {
+            $data['layout'] = $instance['layout'];
+        }
+        else {
+            $data['layout'] = '';
+        }
+
         mybooking_engine_get_template('mybooking-plugin-selector-widget.php', $data); 
     }
  
@@ -52,6 +59,7 @@ class MyBookingRentEngineSelectorWidget extends WP_Widget {
         $sales_channel_code = ! empty( $instance['sales_channel_code'] ) ? $instance['sales_channel_code'] : esc_html__( '', 'text_domain' );
         $family_id = ! empty( $instance['family_id'] ) ? $instance['family_id'] : esc_html__( '', 'text_domain' );
         $rental_location_code = ! empty( $instance['rental_location_code'] ) ? $instance['rental_location_code'] : esc_html__( '', 'text_domain' );
+        $layout = ! empty( $instance['layout'] ) ? $instance['layout'] : esc_html__( '', 'text_domain' );
             ?>
             <p>
                 <label for="<?php echo esc_attr( $this->get_field_id( 'sales_channel_code' ) ); ?>">
@@ -89,6 +97,18 @@ class MyBookingRentEngineSelectorWidget extends WP_Widget {
                     type="text" 
                     value="<?php echo esc_attr( $rental_location_code ); ?>">
             </p>
+            <p>
+                <label for="<?php echo esc_attr( $this->get_field_id( 'layout' ) ); ?>">
+                <?php esc_attr_e( 'layout:', 'text_domain' ); ?>
+                </label> 
+                
+                <select class="widefat"
+                        id="<?php echo esc_attr( $this->get_field_id( 'layout' ) ); ?>"
+                        name="<?php echo esc_attr( $this->get_field_name( 'layout' ) ); ?>">
+                    <option value="horizontal" <?php if ( $layout != 'vertical' ) :?>selected<?php endif; ?>><?php esc_attr_e( 'horizontal', 'text_domain' ); ?></option>   
+                    <option value="vertical" <?php if ( $layout == 'vertical' ) :?>selected<?php endif; ?>><?php esc_attr_e( 'vertical', 'text_domain' ); ?></option>
+                </select>        
+            </p>            
             <?php        
     }
  
@@ -98,6 +118,7 @@ class MyBookingRentEngineSelectorWidget extends WP_Widget {
         $instance['sales_channel_code'] = ( ! empty( $new_instance['sales_channel_code'] ) ) ? strip_tags( $new_instance['sales_channel_code'] ) : '';
         $instance['family_id'] = ( ! empty( $new_instance['family_id'] ) ) ? strip_tags( $new_instance['family_id'] ) : '';
         $instance['rental_location_code'] = ( ! empty( $new_instance['rental_location_code'] ) ) ? strip_tags( $new_instance['rental_location_code'] ) : '';
+        $instance['layout'] = ( ! empty( $new_instance['layout'] ) ) ? strip_tags( $new_instance['layout'] ) : '';
         return $instance;        
     }
 }